+++ src/lib/album_cache.c
@@ -321,7 +321,7 @@ sds album_cache_get_key(sds albumkey, const struct mpd_song *song, const struct 
         }
     }
 
-    // fallback to hashed AlbumArtist::Album::Date tag
+    // fallback to hashed AlbumArtist::Album::<group tag>
     // first try AlbumArtist tag
     albumkey = mpd_client_get_tag_value_string(song, MPD_TAG_ALBUM_ARTIST, albumkey);
     if (sdslen(albumkey) == 0) {
@@ -345,7 +345,7 @@ sds album_cache_get_key(sds albumkey, const struct mpd_song *song, const struct 
     }
     // append album
     albumkey = sdscatfmt(albumkey, "::%s", album_name);
-    //optionally append group tag
+    // optionally append group tag
     if (album_config->group_tag != MPD_TAG_UNKNOWN) {
         const char *group_tag_value = mpd_song_get_tag(song, album_config->group_tag, 0);
         if (group_tag_value != NULL) {