+++ src/mpd_client/search.h
@@ -17,6 +17,7 @@ bool mpd_client_search_add_to_queue(struct t_partition_state *partition_state, c
         unsigned to, enum mpd_position_whence whence, const char *sort, bool sortdesc, sds *error);
 
 bool mpd_client_add_search_sort_param(struct t_partition_state *partition_state, const char *sort, bool sortdesc, bool check_version);
+bool mpd_client_add_search_group_param(struct mpd_connection *conn, enum mpd_tag_type tag);
 sds get_search_expression_album(enum mpd_tag_type tag_albumartist, struct mpd_song *album,
         const struct t_albums_config *album_config);
 sds get_search_expression_album_disc(enum mpd_tag_type tag_albumartist, struct mpd_song *album,