+++ CHANGELOG.md
@@ -46,6 +46,7 @@ It also supports the new "added" timestamp for songs in the mpd database.
 - Feat: Rework data structures to improve thread safety
 - Feat: Improve random number generation
 - Feat: Add option to show audioformat in the footer #1184
+- Upd: Date tag is optional for simple album cache
 - Upd: Improve number conversions and enforcement of limits
 - Upd: Manual add random songs or albums should ignore the jukebox constraints
 - Upd: libmympdclient 1.0.26 (based on libmpdclient 2.23)